Change some stuffs

+++ app/src/main/java/com/teamturtles/greenerme/ViewAccountPage.java
@@ -23,6 +23,7 @@ import com.google.firebase.auth.AuthCredential;
 import com.google.firebase.auth.EmailAuthProvider;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.UserProfileChangeRequest;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -54,6 +55,7 @@ public class ViewAccountPage extends AppCompatActivity implements View.OnClickLi
 
     private EditText oldPassword_EditText, newPassword_EditText, confirmPassword_EditText;
     private EditText newUsername_EditText;
+    private EditText password_EditText;
 
     private TextView username_TextView;
     private String username;
@@ -83,32 +85,6 @@ public class ViewAccountPage extends AppCompatActivity implements View.OnClickLi
     }
 
 
-    private void loadUserInformation() { // added
-
-        if (mAuth.getCurrentUser() == null) {
-            finish();
-            startActivity(new Intent(getApplicationContext(), HomePage_loggedout.class));
-        }
-
-        DatabaseReference mDatabase = databaseReference.child("Users").child(user_id);
-
-        mDatabase.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                if (dataSnapshot.getValue() != null) {
-                    User loggedin_User = dataSnapshot.getValue(User.class);
-                    username = loggedin_User.getUsername();
-                    username_TextView.setText(username);
-                }
-            }
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-                throw databaseError.toException(); // Don't ignore errors
-                // Toast.makeText(getApplicationContext(), "Error!", Toast.LENGTH_SHORT).show();
-            }
-        });
-    }
-
     private void changeUsernamePop() {
 
         mBuilder = new AlertDialog.Builder(this);
@@ -119,8 +95,7 @@ public class ViewAccountPage extends AppCompatActivity implements View.OnClickLi
 
         newUsername_EditText = (EditText) mView.findViewById(R.id.newUsername_EditText);
         username_TextView = (TextView) mView.findViewById(R.id.Username_TextView);
-
-        loadUserInformation();
+        username_TextView.setText(currentUser.getDisplayName());
 
         mBuilder.setView(mView);
         dialog = mBuilder.create();
@@ -144,24 +119,36 @@ public class ViewAccountPage extends AppCompatActivity implements View.OnClickLi
         progressDialog.show();
 
         final String newUsername_final = new_username;
-        final DatabaseReference usernameDatabase = databaseReference.child("Users").child(user_id).child("username");
+        final DatabaseReference usernameDatabase = databaseReference.child("Leaderboard").child(user_id).child("name");
 
-        usernameDatabase.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
+        UserProfileChangeRequest profile = new UserProfileChangeRequest.Builder().setDisplayName(newUsername_final).build();
+        currentUser.updateProfile(profile).addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
             public void onComplete(@NonNull Task<Void> task) {
-
                 progressDialog.dismiss();
 
-                if (task.isSuccessful()) {
-                    usernameDatabase.setValue(newUsername_final);
-                    Toast.makeText(getApplicationContext(),"Username has been edited!", Toast.LENGTH_SHORT).show();
-                    dialog.dismiss();
-                } else {
+                if (!task.isSuccessful()) {
                     Toast.makeText(getApplicationContext(), task.getException().getMessage(), Toast.LENGTH_SHORT).show();
+                } else {
+                    usernameDatabase.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Void> task) {
+
+                            progressDialog.dismiss();
+
+                            if (task.isSuccessful()) {
+                                usernameDatabase.setValue(newUsername_final);
+
+                                Toast.makeText(getApplicationContext(), "Username has been edited!", Toast.LENGTH_SHORT).show();
+                                dialog.dismiss();
+                            } else {
+                                Toast.makeText(getApplicationContext(), task.getException().getMessage(), Toast.LENGTH_SHORT).show();
+                            }
+                        }
+                    });
                 }
             }
         });
-
     }
 
 
@@ -261,6 +248,8 @@ public class ViewAccountPage extends AppCompatActivity implements View.OnClickLi
         deleteAccPop_btn = (Button) mView.findViewById(R.id.deleteAccPop_btn);
         cancel_btn = (Button) mView.findViewById(R.id.cancel_btn);
 
+        password_EditText = (EditText) mView.findViewById(R.id.password_EditText);
+
         mBuilder.setView(mView);
         dialog = mBuilder.create();
         dialog.show();
@@ -272,25 +261,47 @@ public class ViewAccountPage extends AppCompatActivity implements View.OnClickLi
 
     private void deleteAcc() {
 
-        progressDialog. setMessage("Deleting User...");
+        progressDialog.setMessage("Authenticating Password...");
         progressDialog.show();
 
-        currentUser.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+        final String email = currentUser.getEmail();
+        String password = password_EditText.getText().toString().trim();
+
+        AuthCredential credential = EmailAuthProvider.getCredential(email, password);
+
+        currentUser.reauthenticate(credential).addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
             public void onComplete(@NonNull Task<Void> task) {
 
                 progressDialog.dismiss();
 
-                if (task.isSuccessful()) {
-                    Toast.makeText(getApplicationContext(),"User have been deleted!", Toast.LENGTH_SHORT).show();
+                if(task.isSuccessful()){
 
-                    // delete user info from database
-                    FirebaseDatabase.getInstance().getReference().child("Users").child(user_id).removeValue();
-                    finish();
-                    startActivity(new Intent(ViewAccountPage.this, HomePage_loggedout.class));
+                    progressDialog. setMessage("Deleting User...");
+                    progressDialog.show();
+
+                    currentUser.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+                    @Override
+                    public void onComplete(@NonNull Task<Void> task) {
+
+                        progressDialog.dismiss();
+
+                        if (task.isSuccessful()) {
+                            Toast.makeText(getApplicationContext(),"User have been deleted!", Toast.LENGTH_SHORT).show();
+
+                            // delete user info from database
+                            FirebaseDatabase.getInstance().getReference().child("Leaderboard").child(user_id).removeValue();
+                            finish();
+                            startActivity(new Intent(ViewAccountPage.this, HomePage_loggedout.class));
+
+                        } else {
+                            Toast.makeText(getApplicationContext(), task.getException().getMessage(), Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
 
                 } else {
-                    Toast.makeText(getApplicationContext(), task.getException().getMessage(), Toast.LENGTH_SHORT).show();
+                    Toast.makeText(getApplicationContext(),"Authentication Failed", Toast.LENGTH_SHORT).show();
                 }
             }
         });
